refactoring current query/joins to db/legacy.js
