+++ package.json
@@ -7,7 +7,7 @@
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
     "populate-db": "mysql -u root -h booking.cksae9ebsoyz.us-east-2.rds.amazonaws.com -p  < ./schema.sql",
-    "aws-connect": "ssh -i \"Louis-FEC-Booking.pem\" ubuntu@ec2-18-216-104-91.us-east-2.compute.amazonaws.com",
+    "aws-connect": "ssh -i \"somepem.pem\" ubuntu@ec2-18-216-104-91.us-east-2.compute.amazonaws.com",
     "start-pm2": "pm2 start ./server/index.js",
     "list-pm2": "pm2 list",
     "pm2-server": "pm2 show server",
@@ -17,7 +17,7 @@
     "type": "git",
     "url": "git+https://github.com/FangNYC/Booking.git"
   },
-  "author": "louisotter",
+  "author": "louisotter, benmangold",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/FangNYC/Booking/issues"
